requirements.txt
django~=4.1.7
flask
numpy
mysql-connector-python~=8.0.32
